Fix dialog dropdown weird behavior.
If the type of the dropdown values is not defined, the dialog behaves
weirdly. For example it doesn't make the Submit button available if something is
selected and complains something has to be selected. When nothing is touched
when Ordering a catalog using this dialog is permitted. This quite the opposite
to what makes sense.

This is surely a problem of CFME backend -- it shouldn't allow us to not define
a dropdown without a value type defined. But this cannot happen trough
the UI so perhaps not worthy fixing.